+++ src/olmo_core/train/callbacks/console_logger.py
@@ -3,26 +3,11 @@ from dataclasses import dataclass, field
 from fnmatch import fnmatch
 from typing import Dict, List, Optional
 
+from olmo_core.utils import format_float
+
 from .callback import Callback
 
 log = logging.getLogger(__name__)
-
-
-def format_float(value: float) -> str:
-    if value == 0.0:
-        return "0.0"
-    elif value < 0.0001:
-        return f"{value:.2E}"
-    elif value > 1000:
-        return f"{int(value):,d}"
-    elif value > 100:
-        return f"{value:.1f}"
-    elif value > 10:
-        return f"{value:.2f}"
-    elif value > 1:
-        return f"{value:.3f}"
-    else:
-        return f"{value:.4f}"
 
 
 @dataclass